fix(Promotion): PromotionProduct -> Promotion 수정 및 quantity 삭제
